+++ changelog.md
@@ -1,10 +1,11 @@
 ## Sequencer Changelog
 
 ### Version 2.3.0 (V10 only)
-- *Effects* - Deprecated `.belowTokens()`, `.belowTiles()`, and `.aboveLighting()` in favor of `.elevation()` due to fundamental changes in Foundry's V10 update. These methods will be removed in a future update
+- *Effects* - Deprecated `.belowTokens()` and `.belowTiles()` in favor of `.elevation()` due to fundamental changes in Foundry's V10 update. These methods will be removed in a future update
+- *Effects* - Deprecated `.xray()` as it is no longer compatible with Foundry V10's method of rendering
 - *Effects* - Fully removed deprecated methods: `.addPostOverride()`, `.reachTowards()`, `.gridSize()`, `.startPoint()`, `.endPoint()`
 
-### Version 2.2.0 (last V9 update)
+### Version 2.2.0 (last V9 update, except maybe some bug fixes)
 **Additions:**
 - *Effects* - Added `.tieToDocuments()` which allows you to tie an effect to Foundry documents - such as Active Effects or Tokens. When these are deleted, the effect is automatically ended.
 - *Effects* - Added secondary `offset` parameter to `.atLocation()`, `.attachTo()`, `.rotateTowards()`, `.from()`, and `.stretchTo()` which can be used to offset the location of the source or target